// Angular-CRUD/src/app/models/user.model.ts
export class User {
  firstName?: string;
  lastName?: string;
  email?: string;
  mobile?: string;
  weight?: number;
  height?: number;
  bmi?: number;
  bmiResult?: string;
  gender?: string;
  requireTrainer?: string;
  packageName?: string;
  important?: string;
  haveGymBefore?: string;
  enquiryDate?: string;
  id?: number;
}